Fix 32-bit build
Summary:
Fix 32-bit build by using my_add_bigint in place of my_atomic_add64. Do the
same
with my_atomic_cas_bigint. my_atomic_bigint is a 64-bit int on platforms that
support
atomic operations on 64-bit ints.

Test Plan:
mtr, checked for compiler warnings

Reviewed By: ryandmack

+++ storage/innodb_plugin/os/os0file.c
@@ -4471,7 +4471,8 @@ os_io_perf_print(
 		"%lld requests, %lld old, %.2f bytes/r, "
 		"svc: %.2f secs, %.2f msecs/r, %.2f max msecs, "
 		"wait: %.2f secs %.2f msecs/r, %.2f max msecs",
-		perf->requests, perf->old_ios,
+		(long long) perf->requests,
+		(long long) perf->old_ios,
 		perf->bytes / nzero_requests,
 		/* svc: starts here */
 		perf->svc_usecs / 1000000.0,
